Create Rotten Oranges
You are given a 2D grid of size n x m representing the state of oranges in a box. Each cell in the grid can have one of the following three values:

0 → the cell is empty.

1 → the cell has a fresh orange.

2 → the cell has a rotten orange.

Every minute, any fresh orange that is 4-directionally adjacent (up, down, left, or right) to a rotten orange becomes rotten.